creds: replace old cred module with newer
